+++ src/Candoumbe.Pipelines/Components/IReportCoverage.cs
@@ -1,15 +1,13 @@
 ﻿namespace Candoumbe.Pipelines.Components;
 
+using System;
+using System.Linq;
 using Nuke.Common;
 using Nuke.Common.IO;
 using Nuke.Common.Tooling;
 using Nuke.Common.Tools.Codecov;
 using Nuke.Common.Tools.DotNet;
 using Nuke.Common.Tools.ReportGenerator;
-
-using System;
-using System.Linq;
-
 using static Nuke.Common.Tools.Codecov.CodecovTasks;
 using static Nuke.Common.Tools.ReportGenerator.ReportGeneratorTasks;