+++ src/Candoumbe.Pipelines/Components/IUnitTest.cs
@@ -1,16 +1,14 @@
-﻿using Nuke.Common;
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using Nuke.Common;
 using Nuke.Common.CI.AzurePipelines;
 using Nuke.Common.IO;
 using Nuke.Common.ProjectModel;
 using Nuke.Common.Tooling;
 using Nuke.Common.Tools.Coverlet;
 using Nuke.Common.Tools.DotNet;
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-
 using static Nuke.Common.Tools.DotNet.DotNetTasks;
 using static Serilog.Log;